[Feat] 큐레이션 검색 기능을 구현, 기존 검색 VC 리팩토링, 큐레이션 emptyView 설정
