+++ Kindy/Kindy/View Controllers/HomeViewController.swift
@@ -145,8 +145,8 @@ final class HomeViewController: UIViewController {
     
     // 네비게이션 바의 검색 버튼이 눌렸을때 실행되는 함수
     @objc func searchButtonTapped() {
-        let homeSearchViewController = HomeSearchViewController()
-        homeSearchViewController.setupData(items: model.bookstores)
+        let homeSearchViewController = SearchViewController()
+        homeSearchViewController.setupData(items: model.bookstores, itemType: .bookstoreType)
         show(homeSearchViewController, sender: nil)
     }
